Rename everything called 'repository' to 'project'
