+++ src/pages/compare.tsx
@@ -4,13 +4,13 @@ import { AiOutlinePlus } from 'react-icons/ai'
 import { useTrendingProjectsQuery } from '@/generated/gql'
 import Error from '@/components/pure/Error'
 import Loading from '@/components/pure/Loading'
-import TopBar from '@/components/page/repositoryTable/TopBar'
-import Table from '@/components/page/repositoryTable/Table'
-import FilterBar from '@/components/page/repositoryTable/Filterbar'
+import TopBar from '@/components/page/overview/TopBar'
+import Table from '@/components/page/overview/Table'
+import FilterBar from '@/components/page/overview/Filterbar'
 import Chart from '@/components/page/details/Chart'
 import Page from '@/components/side-effects/Page'
 import Button from '@/components/pure/Button'
-import columns from '@/components/pure/RepositoryTable/columns'
+import columns from '@/components/pure/ProjectsTable/columns'
 import { data as chartData } from '@/data/detailPageMocks'
 
 const nullFunc = () => null
